ALTERACAO DO TAMANHO DO CAMPO IDLOTE CONFORME MANUAL DO CONTRIBUINTE AJUSTE NO METODO ClasseParaXmlString PARA GARANTIR QUE A STRING ESTEJA EM UTF8, ALGUNS ESTADOS VALIDAM ESTA FORMATAÇÃO. PEQUENO AJUSTE NA LEITURA DO XML DA NOTA QUANDO V4.00 ESTAVA DANDO ERRO [infNFeSupl.ReadXml]. ATUALIZAÇÃO DOS ENDEREÇOS DE WEBSERVICE DO MATO GROSSO. QUANDO SE TENTA EMITIR UMA NOTA JÁ EMITIDA O RETORNO DO WEBSERVICE VEM [retConsReciNFe] DIFERENTE DO ESPERADO [retEnviNFe], STRING REPLACE NO RETORNO PARA EVITAR EXCEÇÃO. ADICIONADO DELEGATE ValidaCertificado PARA PODER TRATAR O RETORNO DA VALIDAÇÃO DO CERTIFICADO, DECORRENTE DE ALGUM BUG/INCONSISTENCIA QUANDO RODANDO NO IIS NAO CONSEGUI IDENTIFICAR A CAUSA DO ERRO.

+++ NFe.Classes/Servicos/Autorizacao/enviNFe3.cs
@@ -39,7 +39,7 @@ namespace NFe.Classes.Servicos.Autorizacao
     [XmlRoot(ElementName = "enviNFe", Namespace = "http://www.portalfiscal.inf.br/nfe")]
     public class enviNFe3
     {
-        public enviNFe3(string versao, int idLote, IndicadorSincronizacao indSinc, List<NFe> nFe)
+        public enviNFe3(string versao, double idLote, IndicadorSincronizacao indSinc, List<NFe> nFe)
         {
             this.versao = versao;
             this.idLote = idLote;
@@ -59,8 +59,10 @@ namespace NFe.Classes.Servicos.Autorizacao
 
         /// <summary>
         ///     AP03 - Identificador de controle do envio do lote.
+        ///     Número sequencial autoincremental, de controle correspondente ao identificador único do lote enviado.
+        ///     A responsabilidade de gerar e controlar esse número é exclusiva do contribuinte.
         /// </summary>
-        public int idLote { get; set; }
+        public double idLote { get; set; }
 
         /// <summary>
         ///     AP03a - Indicador de Sincronização
